requirements.txt
black==23.1.0
Django==3.2.18
django-environ==0.10.0
django-extensions==3.2.1
django-import-export==3.1.0
djangorestframework==3.14.0
ipython==8.11.0
isort==5.12.0
pre-commit==3.1.1
pytest==7.2.2
pytest-django==4.5.2
